refactor(backend): make the mod file of miscellenous empty
