+++ apps/backend/src/entities/collection.rs
@@ -7,7 +7,7 @@ use sea_orm::entity::prelude::*;
 use sea_orm::ActiveValue;
 use serde::{Deserialize, Serialize};
 
-use crate::miscellaneous::CollectionExtraInformation;
+use crate::models::CollectionExtraInformation;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize, SimpleObject)]
 #[sea_orm(table_name = "collection")]
