+++ apps/backend/src/models.rs
@@ -13,6 +13,7 @@ use database::{
     MediaLot, MediaSource, SeenState, Visibility,
 };
 use derive_more::{Add, AddAssign, Sum};
+use enum_meta::{meta, Meta};
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use schematic::ConfigEnum;
@@ -27,9 +28,87 @@ use strum::Display;
 use crate::{
     entities::{prelude::Workout, user_measurement, workout},
     file_storage::FileStorageService,
-    miscellaneous::CollectionExtraInformation,
     traits::{DatabaseAssetsAsSingleUrl, DatabaseAssetsAsUrls, GraphqlRepresentation},
 };
+
+#[derive(Debug, PartialEq, Eq, Clone, Copy, Serialize, Deserialize, Enum)]
+pub enum CollectionExtraInformationLot {
+    String,
+    Number,
+    Date,
+    DateTime,
+    StringArray,
+}
+
+#[derive(
+    Debug,
+    PartialEq,
+    Eq,
+    Clone,
+    Serialize,
+    Deserialize,
+    SimpleObject,
+    FromJsonQueryResult,
+    InputObject,
+)]
+#[graphql(input_name = "CollectionExtraInformationInput")]
+pub struct CollectionExtraInformation {
+    pub name: String,
+    pub description: String,
+    pub lot: CollectionExtraInformationLot,
+    pub default_value: Option<String>,
+    pub required: Option<bool>,
+}
+
+#[derive(Display, EnumIter)]
+pub enum DefaultCollection {
+    Watchlist,
+    #[strum(serialize = "In Progress")]
+    InProgress,
+    Completed,
+    Monitoring,
+    Custom,
+    Owned,
+    Reminders,
+}
+
+meta! {
+    DefaultCollection, (Option<Vec<CollectionExtraInformation>>, &'static str);
+    Watchlist, (None, "Things I want to watch in the future.");
+    InProgress, (None, "Media items that I am currently watching.");
+    Completed, (None, "Media items that I have completed.");
+    Custom, (None, "Items that I have created manually.");
+    Monitoring, (None, "Items that I am keeping an eye on.");
+    Owned, (Some(
+        vec![
+            CollectionExtraInformation {
+                name: "Owned on".to_string(),
+                description: "When did you get this media?".to_string(),
+                lot: CollectionExtraInformationLot::Date,
+                default_value: None,
+                required: None,
+            }
+        ]
+    ), "Items that I have in my inventory.");
+    Reminders, (Some(
+        vec![
+            CollectionExtraInformation {
+                name: "Reminder".to_string(),
+                description: "When do you want to be reminded?".to_string(),
+                lot: CollectionExtraInformationLot::Date,
+                default_value: None,
+                required: Some(true),
+            },
+            CollectionExtraInformation {
+                name: "Text".to_string(),
+                description: "What do you want to be reminded about?".to_string(),
+                lot: CollectionExtraInformationLot::String,
+                default_value: None,
+                required: Some(true),
+            }
+        ]
+    ), "Items that I want to be reminded about.");
+}
 
 #[derive(Enum, Serialize, Deserialize, Clone, Debug, Copy, PartialEq, Eq)]
 pub enum BackgroundJob {
